refactor(proc): move fns splitting element/int to utils
